Changed reads and writes to do PyBytes vs str
I tried to run this on a server with Python 3.7.3 and received the
following exceptions:

	TypeError: argument 'data': 'str' object cannot be converted to 'PyBytes'

After a little digging it appears in Python 3 the call to open() needs a mode of'rb' vs 'r'.

Found my answer here:
https://stackoverflow.com/questions/1035340/reading-binary-file-and-looping-over-each-byte

Thank you for pkcs7csr!!!!  After changing the "Example" in the
README.rst with the changes here it renewed my certificate like a charm.